requirements.txt
Pillow==4.2.0
rethinkdb
pyyaml
psutil
pendulum
beautifulsoup4
osuapi
-e git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py